requirements.txt
google-cloud-aiplatform 
google-cloud-bigquery 
langchain 
langchain-core 
langchain-text-splitters 
langchain_google_community
langchain-google-vertexai 
langchain-community 
langchain-chroma 
langchain-google-genai 
langchain-openai 
langchain-experimental
faiss-cpu 
pypdf
sqlalchemy
sqlalchemy-bigquery
PyMuPDF 
python-dotenv 
Pillow
wikipedia
pydantic
google-cloud-bigquery-storage
Flask
vertexai